requirements.txt
pandas
numpy
scipy
statsmodels
factor-analyzer
matplotlib
seaborn
scikit-learn
wandb
xgboost
econml
patsy
lightgbm
openpyxl
plotly
basedosdados
google-cloud-storage
PyYAML
streamlit 
google-auth
pandas-gbq
pyarrow